feat: Unify the default website and default site under the same button.
